requirements.txt
python-dotenv==1.1.1
tronpy[mnemonic]==0.5.0
aiohttp==3.12.13